Testing 12: Guest and Payment Preparation
If a user logs in and no clients have been associated with the user raise the:
NoClientsAssignedToUserException.

If a document number cannot be generated because there is no Group associated with it raise the:
GroupException

+++ src/Invoice/Group/GroupService.php
@@ -42,6 +42,11 @@ final class GroupService
      */
     public function deleteGroup(Group $model): void
     {
-       $this->repository->delete($model);
+        // The first three default groups i.e. quote, salesorder, and invoice cannot be deleted
+        if (($model->getName() <> 'Quote Group') 
+                && ($model->getName() <> 'Invoice Group')
+                    && ($model->getName() <> 'Sales Order Group')) {
+            $this->repository->delete($model);
+        }
     }
 }
